+++ iroh-dns-server/Cargo.toml
@@ -31,7 +31,6 @@ humantime-serde = "1.1.1"
 iroh-base = { version = "0.97", path = "../iroh-base", features = ["key"] }
 iroh-metrics = { version = "0.38", features = ["service"] }
 iroh-dns = { version = "0.97", path = "../iroh-dns" }
-iroh-relay = { version = "0.97", path = "../iroh-relay", default-features = false }
 lru = "0.16.3"
 mainline = "6"
 n0-future = "0.3"
