+++ iroh-dns/Cargo.toml
@@ -18,3 +18,4 @@ iroh-base = { version = "0.97.0", path = "../iroh-base", default-features = fals
 n0-error = "0.1"
 n0-future = "0.3"
 simple-dns = "0.9"
+strum = { version = "0.28", features = ["derive"] }
