+++ iroh-relay/Cargo.toml
@@ -41,7 +41,6 @@ n0-error = "0.1"
 n0-future = "0.3"
 num_enum = "0.7"
 pin-project = "1"
-simple-dns = "0.9"
 postcard = { version = "1", default-features = false, features = [
     "alloc",
     "use-std",
